+++ website/src/pages/elements/mod.rs
@@ -1,9 +1,11 @@
 mod button_page;
 mod icon_page;
+mod notification_page;
 mod progress_page;
 mod tag_page;
 
 pub use button_page::ButtonPage;
 pub use icon_page::IconPage;
 pub use progress_page::ProgressPage;
+pub use notification_page::NotificationPage;
 pub use tag_page::TagPage;